refactor(helpers): move duration and locale extensions into core folder
- Moved duration_extension.dart and locale_extension.dart to core directory
- Updated import paths in related files
- Removed obsolete locale_extension_test.dart
