+++ packages/world_countries/lib/src/helpers/typed_locale_delegate.dart
@@ -2,7 +2,7 @@ import "package:flutter/widgets.dart";
 import "package:meta/meta.dart";
 import "package:world_flags/world_flags.dart";
 
-import "../extensions/locale_extension.dart";
+import "../extensions/core/locale_extension.dart";
 import "../extensions/typed_locale_extension.dart";
 import "../models/locale/typed_locale.dart";
 import "../models/typedefs.dart";
