+++ packages/world_countries/lib/src/widgets/pickers/basic_picker.dart
@@ -9,7 +9,7 @@ import "package:world_flags/world_flags.dart";
 
 import "../../constants/ui_constants.dart";
 import "../../extensions/build_context_extension.dart";
-import "../../extensions/duration_extension.dart";
+import "../../extensions/core/duration_extension.dart";
 import "../../extensions/iterable_search_map_extension.dart";
 import "../../extensions/world_countries_build_context_extension.dart";
 import "../../helpers/typed_locale_delegate.dart";
