+++ packages/world_countries/test/src/extensions/core/locale_extension_test.dart
@@ -1,6 +1,6 @@
 import "package:flutter/material.dart";
 import "package:flutter_test/flutter_test.dart";
-import "package:world_countries/src/extensions/locale_extension.dart";
+import "package:world_countries/src/extensions/core/locale_extension.dart";
 import "package:world_countries/src/models/locale/typed_locale.dart";
 import "package:world_flags/world_flags.dart";
